KL-143/rename: change directory
FeedPage 내 components 디렉토리를 생성하였습니다.
FeedPage 단독 컴포넌트를 componets 디렉토리로 이전하였습니다.
import 경로를 수정하였습니다.
