+++ src/pages/FeedPage/FeedPage.jsx
@@ -2,11 +2,11 @@ import React from 'react'
 import { BasicFilterProvider } from '../../contexts/BasicFilterContext'
 import { AdditionalFilterProvider } from '../../contexts/AdditionalFilterContext'
 // import NavBar from '../components/Navbar/NavBar'
-import Thumbnail from './Thumbnail/Thumbnail'
-import BasicFilter from './BasicFilter/BasicFilter'
-import AdditionalFilter from './AdditionalFilter/AdditionalFilter'
-import SelectedField from './SelectedField/SelectedField'
-import ProductList from './ProductList/ProductList'
+import Thumbnail from './components/Thumbnail/Thumbnail'
+import BasicFilter from './components/BasicFilter/BasicFilter'
+import AdditionalFilter from './components/AdditionalFilter/AdditionalFilter'
+import SelectedField from './components/SelectedField/SelectedField'
+import ProductList from './components/ProductList/ProductList'
 // import Footer from '../../components/Footer/Footer'
 import { PageRoot, PageContent, FeedArea } from './FeedPage.style'
