+++ src/pages/FeedPage/components/BasicFilter/BasicFilter.style.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import theme from '../../../styles/theme'
+import theme from '../../../../styles/theme'
 
 const FilterContainer = styled.aside`
   width: 168px;
