+++ src/pages/FeedPage/components/BasicFilter/CitySection.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useCallback, useEffect } from 'react'
 import { Checkbox } from 'antd'
 import PropTypes from 'prop-types'
-import useSelectedCountry from '../../../hooks/useSelectedCountry'
-import useSelectedCity from '../../../hooks/useSelectedCity'
+import useSelectedCountry from '../../../../hooks/useSelectedCountry'
+import useSelectedCity from '../../../../hooks/useSelectedCity'
 import {
   SectionContainer,
   SelectContainer,
