+++ src/pages/FeedPage/components/SelectedField/ResetButton.jsx
@@ -1,11 +1,11 @@
 import React, { useCallback } from 'react'
 import { Button } from 'antd'
 import { GrPowerReset } from 'react-icons/gr'
-import useSelectedCountry from '../../../hooks/useSelectedCountry'
-import useSelectedCity from '../../../hooks/useSelectedCity'
-import useSelectedCategory from '../../../hooks/useSelectedCategory'
-import useSelectedSubCategory from '../../../hooks/useSelectedSubCategory'
-import useSelectedFilter from '../../../hooks/useSelectedFilter'
+import useSelectedCountry from '../../../../hooks/useSelectedCountry'
+import useSelectedCity from '../../../../hooks/useSelectedCity'
+import useSelectedCategory from '../../../../hooks/useSelectedCategory'
+import useSelectedSubCategory from '../../../../hooks/useSelectedSubCategory'
+import useSelectedFilter from '../../../../hooks/useSelectedFilter'
 
 function ResetButton() {
   const { setSelectedCountry } = useSelectedCountry()
