+++ src/pages/FeedPage/components/SelectedField/SelectedField.style.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { StyleSection } from '../FeedPage.style'
+import { StyleSection } from '../../FeedPage.style'
 
 const FieldSection = styled(StyleSection)`
   > div.alignEnd {
