+++ src/pages/FeedPage/components/Thumbnail/Thumbnail.jsx
