+++ src/pages/FeedPage/components/Thumbnail/Thumbnail.style.js
